improv: Use json output of docker instead of bollard
This simplifies tensorman's dependencies, and would work better with
podman.

It seems this results in human readable string rather than more machine
readable formats, but that's perfectly good for our purpose.

The format of tensorman's output is changed somewhat, but probably for
the better, since it better matches docker itself.